// apps/controle-de-horarios-backend/src/controle-horarios/interfaces/globus-horario.interface.ts
export interface IGlobusHorario {
  SETOR_PRINCIPAL_LINHA: string;
  COD_LOCAL_TERMINAL_SEC: number;
  CODIGOLINHA: string;
  NOMELINHA: string;
  COD_DESTINO_LINHA: string;
  LOCAL_DESTINO_LINHA: string;
  FLG_SENTIDO: string;
  DATA_VIAGEM: string;
  DESC_TIPODIA: string;
  HOR_SAIDA: string;
  HOR_CHEGADA: string;
  COD_ORIGEM_VIAGEM: string;
  LOCAL_ORIGEM_VIAGEM: string;
  COD_SERVICO_COMPLETO: string;
  COD_SERVICO_NUMERO: string;
  COD_ATIVIDADE: number;
  NOME_ATIVIDADE: string;
  FLG_TIPO: string;
  COD_MOTORISTA: string;
  NOME_MOTORISTA: string;
  CRACHA_MOTORISTA: string;
  CHAPAFUNC_MOTORISTA: string;
  COD_COBRADOR: string;
  NOME_COBRADOR: string;
  CRACHA_COBRADOR: string;
  CHAPAFUNC_COBRADOR: string;
  TOTAL_HORARIOS: number;
}